+++ src/main/java/com/jasik/momsnaggingapi/domain/user/controller/UserController.java
@@ -1,0 +1,7 @@
+package com.jasik.momsnaggingapi.domain.user.controller;
+
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class UserController {
+}
