+++ src/main/java/com/jasik/momsnaggingapi/domain/user/service/UserService.java
@@ -1,0 +1,13 @@
+package com.jasik.momsnaggingapi.domain.user.service;
+
+import com.jasik.momsnaggingapi.domain.user.repository.UserRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@Transactional(readOnly = true)
+@RequiredArgsConstructor
+public class UserService {
+    private final UserRepository userRepository;
+}
